# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Keel API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Documents",
        "operationId": "list_documents",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentOut"
                  },
                  "type": "array",
                  "title": "Response List Documents"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Upload",
        "operationId": "upload",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Detail",
        "operationId": "detail",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{document_id}/pages/{n}/image": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Page Image",
        "operationId": "page_image",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "n",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{document_id}/fields/{field_id}": {
      "put": {
        "tags": [
          "documents"
        ],
        "summary": "Correct Field",
        "description": "A human correction. It updates the evidence and supersedes any pending approval.",
        "operationId": "correct_field",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Field Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "gate": {
            "type": "string",
            "title": "Gate"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          },
          "proposal": {
            "additionalProperties": true,
            "type": "object",
            "title": "Proposal"
          }
        },
        "type": "object",
        "required": [
          "id",
          "gate",
          "status",
          "summary",
          "proposal"
        ],
        "title": "ApprovalOut"
      },
...
      "Body_upload": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "kind": {
            "type": "string",
            "enum": [
              "order_pad",
              "other",
              "unknown"
            ],
            "title": "Kind",
            "default": "order_pad"
          },
          "source": {
            "type": "string",
            "enum": [
              "upload",
              "camera",
              "email"
            ],
            "title": "Source",
            "default": "upload"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload"
      },
      "Box": {
        "properties": {
          "page_id": {
            "type": "string",
            "title": "Page Id"
          },
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "w": {
            "type": "number",
            "title": "W"
          },
          "h": {
            "type": "number",
            "title": "H"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "page_id",
          "x",
          "y",
          "w",
          "h",
          "text"
        ],
        "title": "Box"
      },
      "CorrectionIn": {
        "properties": {
          "value": {
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "CorrectionIn"
      },
...
      "DocumentDetail": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentOut"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/PageOut"
            },
            "type": "array",
            "title": "Pages"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldOut"
            },
            "type": "array",
            "title": "Fields"
          },
          "checks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Checks"
          },
          "engine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine"
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd"
          },
          "approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "document",
          "pages",
          "fields",
          "checks",
          "engine",
          "cost_usd",
          "approval"
        ],
        "title": "DocumentDetail"
      },
      "DocumentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "duplicate": {
            "type": "boolean",
            "title": "Duplicate",
            "default": false
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "run_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Status"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "kind",
          "status",
          "source",
          "page_count",
          "created_at"
        ],
        "title": "DocumentOut"
      },
...
      "FieldOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "value": {
            "title": "Value"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "engine": {
            "type": "string",
            "title": "Engine"
          },
          "box": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Box"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "path",
          "value",
          "status",
          "confidence",
          "engine",
          "box"
        ],
        "title": "FieldOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "n": {
            "type": "integer",
            "title": "N"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          },
          "has_text_layer": {
            "type": "boolean",
            "title": "Has Text Layer"
          }
        },
        "type": "object",
        "required": [
          "id",
          "n",
          "width",
          "height",
          "has_text_layer"
        ],
        "title": "PageOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}